+++ src/main/java/com/lgtm/easymoney/exceptions/ResourceNotFoundException.java
@@ -1,24 +1,22 @@
-package com.lgtm.easymoney.exceptions.handlers;
-
+package com.lgtm.easymoney.exceptions;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
+import java.io.Serial;
+
 @ResponseStatus(value = HttpStatus.NOT_FOUND)
-// TODO: this FILE is under construction
 // it's intended to handle resource not found exceptions
 // ref: https://4156aseteam.slack.com/archives/C046MNCJ2U9/p1665800409627119?thread_ts=1665794183.617079&cid=C046MNCJ2U9
 public class ResourceNotFoundException extends RuntimeException {
+    @Serial
     private static final long serialVersionUID = 1L;
 
-    private transient ResponseEntity apiResponse;
-
-    private String resourceName;
-    private String fieldName;
-    private Object fieldValue;
+    private final String resourceName;
+    private final String fieldName;
+    private final Object fieldValue;
 
     public ResourceNotFoundException(String resourceName, String fieldName, Object fieldValue) {
-        super();
+        super(String.format("%s not found with %s: '%s'", resourceName, fieldName, fieldValue));
         this.resourceName = resourceName;
         this.fieldName = fieldName;
         this.fieldValue = fieldValue;
@@ -34,15 +32,5 @@ public class ResourceNotFoundException extends RuntimeException {
 
     public Object getFieldValue() {
         return fieldValue;
-    }
-
-    public ResponseEntity getApiResponse() {
-        return apiResponse;
-    }
-
-    private void setApiResponse() {
-        String message = String.format("%s not found with %s: '%s'", resourceName, fieldName, fieldValue);
-
-//        apiResponse = new ResponseEntity(Boolean.FALSE, message);
     }
 }
